* 'devel' of https://github.com/arangodb/arangodb: (31 commits)
  bug-fix/internal-issue-#591 (#9639)
  miscellaneous adjustments to pacify cppcheck (#9677)
  Bug fix/simon 19 08 07 (#9655)
  add a way to output the version in the tests (#9673)
  removed unused variable
  hide MMFiles-specific information when we don't need it (#9667)
  remove unused instance variable (#9662)
  Add missing includes for Windows builds. (#9657)
  Set jsunity output to gtest style (#9646)
  add the ability to have cluster special examples (#9613)
  Create TakeoverShardLeader job. (#9653)
  Aardvark: Add k Shortest Paths example graph to UI (#9491)
  [devel] Resign Leadership (#9427)
  get rid of _resolvedNames and _nameLock (#9649)
  Add constructors to fix cppcheck warnings. (#9648)
  Bug fix/optimizations 06 08 2019 (#9641)
  fix compile error
  Decentralize includes (#9623)
  add error to failure (#9637)
  Fast Controlled Leaderchange (#9608)
  ...

+++ arangod/Scheduler/SupervisedScheduler.cpp
@@ -22,6 +22,9 @@
 /// @author Achim Brandt
 ////////////////////////////////////////////////////////////////////////////////
 
+#include <velocypack/Value.h>
+#include <velocypack/velocypack-aliases.h>
+
 #include "SupervisedScheduler.h"
 #include "Scheduler.h"
 
@@ -32,7 +35,9 @@
 #include "Cluster/ServerState.h"
 #include "GeneralServer/Acceptor.h"
 #include "GeneralServer/RestHandler.h"
+#include "Logger/LogMacros.h"
 #include "Logger/Logger.h"
+#include "Logger/LoggerStream.h"
 #include "Random/RandomGenerator.h"
 #include "Rest/GeneralResponse.h"
 #include "Statistics/RequestStatistics.h"
@@ -322,9 +327,9 @@ void SupervisedScheduler::runWorker() {
     state->_sleepTimeout_ms = 1000;
   }
 
-  if (id < 32) {
+  if (id < 32U) {
     // 512 >> 32 => undefined behavior
-    state->_queueRetryCount = (512 >> id) + 3;
+    state->_queueRetryCount = (uint64_t(512) >> id) + 3;
   } else {
     // we want at least 3 retries
     state->_queueRetryCount = 3;
